chore: Update wizard form to include option for cloning workshops repository
The code changes in the `index.js` file of the `wizard` page add a checkbox option for cloning the workshops repository. When the checkbox is selected, the workshops repository will be cloned during the institute creation process. This provides flexibility for users who want to include workshops in their institute.