chore: Include GitHub token expiry in local storage
This commit adds the `githubTokenExpiry` key to the local storage, storing the expiry time of the GitHub token. The expiry time is set to one day from the current date. This ensures that the token is removed from the local storage after it expires, improving the security of the application.