chore: Update Institute page to include contact information
The code changes in the `index.js` file of the `inst` page update the contact information section. The `<p>` tags are replaced with `<h4>` for the contact name and an email link is added for the email address. This change improves the readability and user experience of the Institute page.