refactor: Update JupyterLoad.js to use HTTPS for finalURL
This commit updates the JupyterLoad.js file to use HTTPS for the finalURL variable. Previously, it was using an insecure HTTP URL, which could lead to security vulnerabilities. By switching to HTTPS, we ensure a secure connection when loading the Jupyter notebook. This change improves the overall security of the application.