refactor: Add support for 'command_line' and 'html_css' languages in DrawerEditor
This commit updates the DrawerEditor component to add support for the 'command_line' and 'html_css' languages. Previously, the component only supported 'computer' and 'html' languages. By including these additional languages, we enhance the versatility and functionality of the component, allowing users to work with a wider range of programming languages. This change improves the overall user experience and flexibility of the application.